Conflicts:
	src/java/servlets/LoginServlet.java
	web/views/aside.jsp
	web/views/cv.jsp

+++ src/main/java/com/mii/cvonlinerestserver/dao/LowonganDAO.java
@@ -26,7 +26,7 @@ public class LowonganDAO {
 		return lowonganRepository.findAll();
 	}
 	
-	public Lowongan findOne(Long id) {
+	public Lowongan findOne(Integer id) {
 		return lowonganRepository.findOne(id);
 	}
